% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ML_completeness.R
\name{extract_ML_completeness}
\alias{extract_ML_completeness}
\title{Fits the Maximum Likelihood method to estimate completeness of the fossil record
developed by Foote.}
\usage{
extract_ML_completeness(fr)
}
\arguments{
\item{fr}{a matrix, data.frame, or tbl with 2 columns of per-taxon first and
last occurrence, relative to the modern. If it's a list, it'll try to
extract the matrix from the second entry.}
}
\value{
a numeric vector with the DR estimated for each tip
}
\description{
Fits the Maximum Likelihood method to estimate completeness of the fossil record
developed by Foote.
}
\details{
Runs the analysis of the fossil record completeness developed by:
}
\examples{
set.seed(666)
record = paleotree::simFossilRecord(
p = 0.3,
q = 0.1,
nruns = 1,
r = 3,
nTotalTaxa = c(30, 40)
)
firstlast = sapply(record, function(x)
range(x[["sampling.times"]]))
FR = t(apply(firstlast, MARGIN = 2, function(x)
ifelse(is.finite(x), x, NA)))
res = extract_ML_completeness(FR)
res$par
}
\references{
Foote, M. 1997. Paleobiology 23(3):278–300.
Foote, M., and D. M. Raup. 1996. Paleobiology 22(2):121–140.
}
\author{
Mauro TC Sugawara, based on David W. Bapst make_durationFreq* functions
 from paleotree
}
